Rename Dao to UnitDao
As each Dao (e.g., CourseDao) is for a specific model. Changing it's
name to UnitDao might make more sense. There should be a Dao that serves
the whole system (e.g., adding a lecturer to a module), which seperate
the groups of functionalties. It may be better this way to improve the
maintainability, tho the JDBC & sql queries are inevitably hard to maintain.