Refactor: Change all Date to LocalDate
Date is hard to use and maintain. The date stored in DBMS is simply
a YYYY-MM-DD string, it should ignore timezone and time. The timezone
that the system is at will affect who the date in the DBMS is converted
to the Date object in java, which in turn causing the conversion unreliable.

E.g., in the java, a "2020-01-21" can be converted to "2020-01-20" in the DBMS,
as the timezone that the java program used can be different from the one used by
the DBMS. This is unfortunately, implementation dependent.

All the Date impl is now changed to LocalDate. The conversion between the date
in DBMS and LocalDate is no longer conducted by the JDBC through java.sql.Date.
Instead, such conversion is achieved through parsing the string, which is very
consistent.

E.g., "2020-01-21" is converted to 2020-01-21 regardless of timezone.