Add CommonLvController
As the commonLv itself requires a lots of methods and event handlers.
All methods for the commonLv are moved to the new class CommonLvController.
This way, the code is more readable, and methods are better organised.

Additionally, the event handler for double click on the commonLv is
implemented, that displays the content of the selected item.